changed int to long

+++ interceptor.c
@@ -252,6 +252,7 @@ void (*orig_exit_group)(int);
 void my_exit_group(int status) {
 
 	spin_lock(&pidlist_lock);
+	//curent might be null
 	del_pid(current->pid);
 	spin_unlock(&pidlist_lock);
 	orig_exit_group(status);
@@ -278,7 +279,7 @@ void my_exit_group(int status) {
  */
 asmlinkage long interceptor(struct pt_regs reg) {
 
-	int syscall = reg.ax;
+	long syscall = reg.ax;
 	int monitoredFlag = table[syscall].monitored;
 	//check if the syscall has any pid monitored, if yes log message
 	if ((monitoredFlag == 2) || (monitoredFlag == 1 && check_pid_monitored(syscall, current->pid) == 1)){
